fix!: Use proper OR operator instead of XOR
BREAKING CHANGE - this can break previous behaviour and should be marked
as a breaking change.

This is breaking because

```js
let aWriteInterval = 10;

aWriteInterval |  10 * 1000; // 10010
aWriteInterval || 10 * 1000; // 10
```

You obviously meant to allow changing the write interval instead of
XOR'ing it with a default `10000` one, right?:D

I noticed this just now - I'm trying to automate the openapi docs
generation for production.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>