Create `onWriteCallback` functionality
I want to do some modifications everytime the file is updated, so
allowing a user to create a callback once the file is written is very
useful.

I tried writing tests for this but couldn't figure out how - maybe one
should create a file inside the callback and in the tests check if the
file was created to indicate that stuff works. After that - delete the
temp file.

---

I think we could also allow something like `preWrite` callback,
providing the spec that would be written.

This brings attention to another issue - we should definitely start
using an object for all the options, because this is already kinda
bloated, and it could get even more.

See https://github.com/mpashkovskiy/express-oas-generator/issues/35

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>